requirements.txt
streamlit>=1.30.0
easyocr>=1.7.1
opencv-python-headless>=4.9.0.80
numpy>=1.24.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0